retain exact functionality, but split the files apart
gotta compartmentalize, otherwise I won't be able to think straight
