+++ compiler/generator/filter/scopes.go
@@ -1,0 +1,5 @@
+package filter
+
+type frugalFilterScopesSpec struct {
+	All *bool `yaml:"all"`
+}
