+++ compiler/generator/filter/services.go
@@ -1,64 +1,10 @@
-package generator
+package filter
 
 import (
 	"strings"
 
 	"github.com/Workiva/frugal/compiler/parser"
 )
-
-type frugalFilterYaml struct {
-	Included *frugalFilterSpec `yaml:"included"`
-	Excluded *frugalFilterSpec `yaml:"excluded"`
-}
-
-type frugalFilterSpec struct {
-	Services []frugalFilterServiceSpec `yaml:"services"`
-	Structs  []filterFrugalStruct      `yaml:"structs"`
-	Scopes   *frugalFilterScopesSpec   `yaml:"scopes"`
-}
-
-func (ffs *frugalFilterSpec) isServiceSpecified(
-	s *parser.Service,
-) bool {
-	if ffs == nil {
-		return false
-	}
-
-	for _, fs := range ffs.Services {
-		if fs.isService(s) {
-			return true
-		}
-	}
-	return false
-}
-
-func (ffs *frugalFilterSpec) isEntireService(
-	s *parser.Service,
-) bool {
-	if ffs == nil {
-		return false
-	}
-
-	for _, fs := range ffs.Services {
-		if fs.isService(s) && fs.Entire != nil {
-			return *fs.Entire
-		}
-	}
-
-	return false
-}
-
-func (ffs *frugalFilterSpec) shouldRemoveScope(
-	s *parser.Scope,
-) bool {
-	if ffs == nil || ffs.Scopes == nil {
-		return false
-	}
-
-	// Currently, we don't have the ability to filter at a per-scope level.
-	// It's all or nothing.
-	return ffs.Scopes.All != nil && *ffs.Scopes.All
-}
 
 type frugalFilterServiceSpec struct {
 	Name    string   `yaml:"name"`
@@ -79,28 +25,6 @@ func (ffs *frugalFilterServiceSpec) hasMethod(
 		}
 	}
 	return false
-}
-
-type frugalFilterScopesSpec struct {
-	All *bool `yaml:"all"`
-}
-
-type filterFrugalStruct struct {
-	Name string `yaml:"name"`
-}
-
-func shouldEntirelyRemoveService(
-	gf *frugalFilterYaml,
-	s *parser.Service,
-) bool {
-	return gf.Excluded.isEntireService(s)
-}
-
-func shouldEntirelyRemoveScope(
-	gf *frugalFilterYaml,
-	s *parser.Scope,
-) bool {
-	return gf.Excluded.shouldRemoveScope(s)
 }
 
 func applyFilterToService(
