+++ compiler/generator/filter/structs.go
@@ -1,0 +1,5 @@
+package filter
+
+type filterFrugalStruct struct {
+	Name string `yaml:"name"`
+}
